Fix video stall on playback resume.
Allocation query can stall the straming thread when sent with sink
waiting in pre-roll. Effectively blocking the processing of
data.

This change implements dropping of the query in source pad, since we
don't use proposed allocation anywhay. And also removes the extra
buffering target for video (reducing the time to resume the playback).

Change-Id: I32b4a303763e958ef8fcadebd9f2e4820594be24
Signed-off-by: Eugene Mutavchi <Ievgen_Mutavchi@comcast.com>