Bring back extra buffering time for video
In some cases (High bitrate tests) seeking close to end of segment may
require longer time to demux and provide video to the playback
pipeline.

We're dropping allocation queries, so nothing should be blocking app
src streaming thread and it should be safe to wait for a bit longer
for video to buffer.

Lets allocate more time for video buffering.

Change-Id: Ie57cfce1d24b8a68ec5adeae4fdd5f8782307f7f